SPDB-381: tests: fix tests that use custom comparators
After SPDB-197, custom comparators that don't return false for the check
CanKeysWithDifferentByteContentsBeEqual() error out on DB open due to the
automatic addition of a filter policy on our part. However, most of these
tests indeed never compare keys with different byte contents as equals, so
the easy fix is to just override CanKeysWithDifferentByteContentsBeEqual()
to return false.

The tests that are failing because they do compare keys with different
byte contents as equal need to override their filter policy, as done in
SPDB-197.

While at it, also fix an embarrassing bug in the filter policy implementation
done as part of SPDB-197.