optimize memtable switch latency
With the new speedb memtable, each memtable switch needs to wait
for the allocation and initialization of ~8MiB, which creates
spikes in latency when switching memtables. Add a thread per CF
to allocate new memtables in the background so that they'll be
ready to be used when a switch happens.
In case memtable is not ready when switch occor the memtable will be created
immediately

pending memtable shouldnt be calculate as used