Avoid flooding disk IO on file purges
When we need to purge a large amount of files, or large files
which don't have their blocks allocated contiguously, or on
bitmap-based filesystems (even when we have a contiguous range of blocks for each file),
we may incur a high load on the disk when purging obsolete files
that would eat into the bandwidth available to user operations and compactions.
To avoid this situation, this change truncates files into modestly sized chunks
and then deletes each one of them, instead of deleting the entire file in one go.

Signed-off-by: Amnon Hanuhov <AmnonSWE@gmail.com>