cmake: use the old plugin infra and add support for *_FUNC registration (#80)
The new Customizable plugin infra in CMake relies on regex parsing of
Makefiles, which is very brittle and clunky, as well as being unintuitive
for CMake users.

The only thing missing from the old infra is plugin `*_FUNC` registration,
so just add support for that to the old infra, and remove the Makefile
parsing logic.