crash_test: fix incorrect passing of the memtablerep argument (#93)
As part of #90, I accidentally added a comma at the end of the
line when choosing the memtable reperesntation to use. This is
a tuple construction syntax and was formatted as such by Black,
but unfortunately missed by me and during review.

Fix it by removing the comma so that the argument isn't passed
to db_stress as a Python tuple.