crash_test: remove the hash-based memtable from crash test (#98)
The test seems to fail for now, so in order to avoid headaches for devs
working on `main`, let's not test the new memtable until the issues are
fixed. This is not intended as a fix for #98, but rather as a temporary
workaround to allow testing to continue, since no one is supposed to be
using the new memtable yet.