licence: remove the GPL as an alternative licence (#118)
We only license Speedb under the Apache 2.0 licence, so we need to remove
the GPL text from the repository. Since GNU Parallel is currently
vendored in the repo and requires having a copy of the GPL text, remove
it as well and depend on the system installed binary if exists (out build
instructions ask users to install it anyway, and it's installed in CI).