build: fix the java test target dependencies (#128)
The test target in the Java Makefile has an incorrect dependency graph
that could cause tests to start executing before the compilation finished
(or even started) when make is invoked in job server mode.

Fix it by specifying target dependencies such that running the tests depends
on having the tests compiled first.