makefile: avoid setting TEST_TMPDIR unnecessarily (#140)
The changes in #119 included an upstream commit that moved the TEST_TMPDIR
setting to common.mk and executed it unconditionally.

This is causing pollution of /dev/shm unnecessarily, and also causes the
crash tests and other targets to use /dev/shm when they originally didn't.
On machines with small /dev/shm this leads to the crash tests failing with
an out of space error.

Make the setting of TEST_TMPDIR and the creation of a temporary directory
under it execute only once during a single make invocation and only for
the check target, as it was originally. This requires dealing with make
restarts (due to makefiles getting remade) in the form of adding another
temporary configuration file named test_config.mk