build: replace usages of egrep with grep -E (#166)
egrep and fgrep have been deprecated for a long time in GNU grep, and as
of release 3.8 using them prints a warning message, as stated in the 3.8
release notes:

> The egrep and fgrep commands, which have been deprecated since
> release 2.5.3 (2007), now warn that they are obsolescent and should
> be replaced by grep -E and grep -F.

egrep is used during the build to query the version, and with the warning
that's now being printed this adds unneeded noise to the output of the
build. Replace the usage of egrep with grep -E to get rid of the warning.