db_impl_open: make sure not to create a directory if it's missing
Currently the DB creates a directory and opens a database even if the
`create_if_missing` option is set to `false`. This is due to the fact
that the logger creates the directory unconditionally and the sequence of
operations for opening the DB isn't careful to avoid making changes to
the filesystem before checking `create_if_missing`.