deletefile_test: fix breakage caused by the compaction threads change (#217)
In #194 the default value for background compaction threads was changed
to 8 (from 1). This caused some tests to fail and fixes were implemented
as part of #197. Alas, it seems that the fixes weren't complete, as under
stress a test still fail.

Make it so the test is truly fixed now. This can be verified by running
them in a loop with the machine overloaded:

```
$ while ./deletefile_test --gtest_filter=DeleteFileTest.BackgroundPurgeCFDropTest; do sleep 1; done
```