SPDB-347: Add a version number to our builds
The change set includes:
1) A header file that contains the Major, Minor, and Patch versions of speedb
as a macro, as well as a set of functions returning info about how/when/where
this version of speedb was created.

also includes all changes done on build_version.cc.in in the following commits:
1. version: remove superfluous build property
The `speedb_build_spdb_key` property is unused and was accidentally
imported as part of #1.

2.general: replace RocksDB references in strings with Speedb (#64)
This includes references in statuses as well as tools output.