Dynamic Delay: new feature - Dynamic Delay (#276)
Calculate the delayed write rate based on the current state of the CF
based on the value of delayed_write_rate that the user initially set.
The calculation reduces speed linearly with regards to how much
the CF has exceeded the slowdown conditions (L0 files and pending bytes).
The slowdown for writing to the last memtable (when theres more than 3)
is 10 fold