unit tests: fix db_test PurgeInfoLogs by creating unique db path (#273)
create a unique db path for each new db opened in a suite.
the PerThreadDBPath creates a db path based on the PID and the thread so
that every test in the suite (e.g. db_test) has the exact same db path.
this leads to unwanted leftovers which caused an error in PurgeInfoLogs.