unit tests: fix CompactionServiceTest.RemoteEventListener (#314)
the test assumes that each compaction job_id from the remote compaction
is unique but in fact all the job_ids are 1.
the test passed thus far since there was just 1 bg compaction by default.
once we increased it to 8, there could be concurrent remote compactions
which led to this error.