Dynamic Delay: CFs wont override each other (#358)
Currently, in a setting of multiple cfs, the delayed_write_rate_ value is overridden by the latest
cf that entered a delay state. so if a cf registered a delay state that
is less severe than the previos cf, it will override the previous value of
delayed_write_rate_ with a higher rate than the system should use.

fix it by storing the delay requirements of all the CFs in the write controller
and only set a new delayed_write_rate_ when its smaller than the min.