Global Write Controller: share write controller across dbs (#346)
enable sharing the Write Controller across dbs through the db options
while still retaining the old behavior.
when shared between dbs, the write controller keeps track of all the delay
requirements of all cfs in the dbs and enforces on all of them the
lowest write rate required.

NOTE: plz note that setting a write controller in the options is only
valid with use_dynamic_delay = true.

Dynamic Delay: CFs wont override each other (#358)

Currently, in a setting of multiple cfs, the delayed_write_rate_ value is overridden by the latest
cf that entered a delay state. so if a cf registered a delay state that
is less severe than the previous cf, it will override the previous value of
delayed_write_rate_ with a higher rate than the system should use.

fix it by storing the delay requirements of all the CFs in the write controller
and only set a new delayed_write_rate_ when its smaller than the min.