unit tests: Fix GlobalAndWBMSetupDelay
The problem was that calling Flush with FlushOptions.wait = true,
doesn't wait for the memory to be free which also happens during the
flush process.
So the fix is adding a sync point which delays the assertion of checking
the WBM memory usage to after the flush cleared the memtable memory.