options: Add generic validation implementation for use cases
A UseCase class is a Configurable class which represents a specific use case
in rocksdb/speedb which we can use to populate options according to the
specification of that use case, and/or validate the user's options against
the expected possible configurations which fit that use case.

As such, we add here 2 methods in Configurable which essentially take
a UseCase and a set of options, and validates that the values
of these options matches the values that the use case requires.