Add ability to print non-standard options to log and other files
Replace the "GetPrintableOptions" with "SerializePrintableOptions".  Now options can be added to the Properties in name/value format and outputted to a variety of formats (via the OptionsFormatter).

A new ConfigOptions::Depth was added to tell that these "printable" options should also be included in the output.

The XXOptions::Dump methods were changed to use these new methods.  The Configurable::GetPrintableOptions method was changed from virtual to non and also uses this functionality.