Change Configurable::SerializeOptions method to return a map (#619)
Changes the Configurable::SerializeOptions method to return a map of options. This map is equivalent to the input map for the ConfigureOptions method. By returning a map, it will be possible in the future to add different output representations (such as DUMP, JSON, Properties) formats to the string representations of the configurable objects.

This PR is a first-step to a few additional changes. For example, is a gateway to:

Only serializing changed options;
Allowing different string formats for options (e.g. Default, Properties, Log/Dump, JSON, etc)