Rebase 8.6.7: Disable failing tests in db_bloom_filter_test.cc
As part of the rebase on RocksDB 8.6.7, the following commit:
39f5846 - Much better stats for seeks and prefix filtering (#11460)

Included changes that cause speedb_db_bloom_filter_test.cc to fail some tests.

speedb_db_bloom_filter_test.cc was essentially a duplicate of db_bloom_filter_test.cc with focus on paired bloom filter.

It might be worth doing away with speedb_db_bloom_filter_test.cc altogether, and integrating the paired bloom filter tests in the existing db_bloom_filter_test.cc. This will avoid the duplication of a lot of existing testing code and also solve most of the problems. The reason for not doing so in the first place were to keep the paired bloom filter as an external plug in. However, this seems not to make sense considering the current state.